requirements.txt
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2
mlconfig==0.1.7
mlflow==1.28.0
Pillow==8.4.0
opencv-python==4.6.0.66
numpy==1.26.4
imgaug==0.4.0
tqdm==4.64.0
pyyaml==6.0.2
requests==2.32.3
entrypoints==0.4
llvmdev==19.1.7